+++ src/log.ts
@@ -1,4 +1,4 @@
-import * as debug from 'debug';
+import debug from 'debug';
 
 export const log = (type, ...args) => {
   const types = {