+++ src/util/api.ts
@@ -1,35 +1,5 @@
 import {lodash as _} from './lodash';
 
-/**
- * Determine if a value is a boolean representation.
- * @param value
- * @return {boolean}
- */
-export const isBoolean = (value) => {
-  if (typeof value === 'boolean') {
-    return true;
-  } else if (typeof value === 'string') {
-    value = value.toLowerCase();
-    return (value === 'true') || (value === 'false');
-  }
-  return false;
-};
-
-/**
- * Quick boolean coercer.
- * @param value
- * @return {boolean}
- */
-export const getBoolean = (value) => {
-  if (typeof value === 'boolean') {
-    return value;
-  }
-  if (typeof value === 'string') {
-    return (value.toLowerCase() === 'true');
-  }
-  return !!value;
-};
-
 export const eachValue = (components, data, fn, context, path = '') => {
   const promises = [];