+++ src/util/index.ts
@@ -1,9 +1,11 @@
 import * as api from './api';
+import {fetch} from './fetch';
 import {formio} from './formio';
 import {lodash} from './lodash';
 
 export const util = {
   api,
+  fetch,
   formio,
   lodash,
 };