+++ src/util/lodash.ts
@@ -1,17 +1,17 @@
-import * as camelCase from 'lodash/camelCase';
-import * as capitalize from 'lodash/capitalize';
-import * as chunk from 'lodash/chunk';
-import * as find from 'lodash/find';
-import * as get from 'lodash/get';
-import * as has from 'lodash/has';
-import * as isEmpty from 'lodash/isEmpty';
-import * as isEqual from 'lodash/isEqual';
-import * as isNumber from 'lodash/isNumber';
-import * as isString from 'lodash/isString';
-import * as omit from 'lodash/omit';
-import * as set from 'lodash/set';
-import * as unset from 'lodash/unset';
-import * as update from 'lodash/update';
+import camelCase from 'lodash/camelCase';
+import capitalize from 'lodash/capitalize';
+import chunk from 'lodash/chunk';
+import find from 'lodash/find';
+import get from 'lodash/get';
+import has from 'lodash/has';
+import isEmpty from 'lodash/isEmpty';
+import isEqual from 'lodash/isEqual';
+import isNumber from 'lodash/isNumber';
+import isString from 'lodash/isString';
+import omit from 'lodash/omit';
+import set from 'lodash/set';
+import unset from 'lodash/unset';
+import update from 'lodash/update';
 
 /**
  * This is a wrapper around lodash functions. It allows us to optimize builds and not import all of lodash.