+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.6.8
+### Fixed
+ - Syntax of Authorization Bearer tokens.
+
 ## 0.6.7
 ### Added
  - Fetch library to replace request.js.